fix: 측정 단위 수정

See merge request s11-final/S11P31B108!72

+++ backend/src/main/java/sleepGuardian/domain/totalInformation/dto/SleepSolutionResponseDTO.java
@@ -9,9 +9,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class SleepSolutionResponseDTO {
     private int totalInformationId;
-    private double illumination;
-    private double noise;
-    private double humidity;
-    private double temperature;
+    private int illumination;
+    private int noise;
+    private int humidity;
+    private int temperature;
     private String solution;
 }
